Add winning sequence check. Add user input to human player. Add mcts configuration and condition for building tree. Refactor modules and files names.
